+++ R/render.R
@@ -7,6 +7,8 @@
 #' the duration of the module is calculated and used for rendering.
 #' @param options A list of options used for rendering the audio. Use
 #' [`pt2_render_options()`] to obtain default options, or modify them.
+#' @param position Starting position in the pattern sequence table (`pt2_pattern_table()`).
+#' Should be a non negative value smaller than the mule length (`pt2_length()`).
 #' @param ... Ignored
 #' @returns Rendered audio inheriting the [`audio::audioSample()`] class.
 #' @examples
@@ -21,8 +23,8 @@ pt2_render <- function(x, duration = NA, options = pt2_render_options(), ...) {
 #' @rdname pt2_render
 #' @name pt2_render
 #' @export
-pt2_render.pt2mod <- function(x, duration = NA, options = pt2_render_options(), ...) {
-  render_mod_(x, as.numeric(duration), options) |>
+pt2_render.pt2mod <- function(x, duration = NA, options = pt2_render_options(), position = 0L, ...) {
+  render_mod_(x, as.numeric(duration), options, as.integer(position)) |>
     matrix(nrow = 2, byrow = FALSE) |>
     audio::audioSample(rate = options$sample_rate)
 }
@@ -54,8 +56,8 @@ NULL
 #' ctrl <- play(mod)
 #' }
 #' @method play pt2mod
-play.pt2mod <- function(x, duration = NA, options = pt2_render_options(), ...) {
-  x <- pt2_render(x, duration = duration, options = options, ...)
+play.pt2mod <- function(x, duration = NA, options = pt2_render_options(), position = 0L, ...) {
+  x <- pt2_render(x, duration = duration, options = options, position = position, ...)
   rate <- attributes(x)$rate
   audio::play.audioSample(x, rate = rate)
 }
@@ -87,4 +89,26 @@ pt2_render_options <- function(...) {
     result[[i]] <- repl[[i]]
   }
   return (result)
+}
+
+#' Calculate the duration of the module
+#' 
+#' How long a module will play depends on several aspects such as
+#' the length of the pattern sequence table (`pt2_pattern_table()`, `pt2_length()`),
+#' the speed and tempo at which the patterns are defined,
+#' loops, pattern breaks and delay effects. The duration in seconds of the
+#' module is calculated by this function.
+#' @inheritParams pt2_render
+#' @param x Object for which to determine the duration. Should be of class `pt2mod`.
+#' @returns The duration in seconds (as a `difftime` class object)
+#' @author Pepijn de Vries
+#' @examples
+#' mod <- pt2_read_mod(system.file("mod.intro", package = "ProTrackR2"))
+#' 
+#' pt2_duration(mod)
+#' @export
+pt2_duration <- function(x, options = pt2_render_options(), position = 0L, ...) {
+  dur <- mod_duration(x, options, as.integer(position)) |>
+    as.difftime(units = "secs")
+  return(dur)
 }