+++ NEWS.md
@@ -1,6 +1,8 @@
-ProTrackR2 v0.0.1.0001
+ProTrackR2 v0.0.2
 -------------
 
+  * Calculate song duration
+  * Start at module positions other than 0
   * Format patterns conform OpenMPT specs
 
 ProTrackR2 v0.0.1